+++ profiling/Main.hs
@@ -25,9 +25,9 @@ main =
           where
             host = "localhost"
             port = 5432
-            user = "postgres"
-            password = ""
-            database = "postgres"
+            user = "hasql_test"
+            password = "hasql_test"
+            database = "hasql_test"
 
 
 -- * Sessions