+++ tasty/Main/Connection.hs
@@ -19,9 +19,9 @@ with handler =
           where
             host = "localhost"
             port = 5432
-            user = "postgres"
-            password = ""
-            database = "postgres"
+            user = "hasql_test"
+            password = "hasql_test"
+            database = "hasql_test"
     use connection =
       lift $ handler connection
     release connection =