+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.5.0.2
+
+- Add `exceptions` instances (`MonadThrow`, `MonadCatch`, `MonadMask`) to the
+  `Session` monad.
+
 # 1.5
 
 - Added column number to `RowError` (breaking change).